requirements.txt
# LLM e Agentes
langchain>=0.1.0
langchain-groq>=0.0.1
langgraph>=0.0.20

# Banco de dados
sqlalchemy>=2.0.0

# Manipulação de dados
pandas>=2.0.0
numpy>=1.24.0

# Gráficos
matplotlib>=3.7.0

# Busca de notícias
duckduckgo-search>=4.0.0

# Geração de PDF
reportlab>=4.0.0

# Utilitários
python-dotenv>=1.0.0

# Desenvolvimento (opcional)
jupyter>=1.0.0
ipykernel>=6.0.0